Implemented graph initialization in the examples project and implemented adjacency lists for graph traversal

+++ C++ Library/PathfindingLib/PathfindingLib/PGraph.h
@@ -6,7 +6,7 @@
 class PGraph
 {
 public:
-	PGraph(blaze::StaticVector<double, 3UL> origin, blaze::StaticVector<int, 3UL> dimensions, double scale);
+	PGraph(blaze::StaticVector<double, 3UL> origin, blaze::StaticVector<double, 3UL> dimensions, double scale);
 	~PGraph();
 };
